requirements.txt
python-nmap
matplotlib
networkx
mac-vendor-lookup
